+++ src/logica/Memoria.java
@@ -9,10 +9,10 @@ import interfaces.IRegistro;
 
 public class Memoria {
     
-    private int capacidadMemoria; // capacidad de la memoria en bytes
+    private int capacidadMemoria; // capacidad de la memoria en ints
     
-    // Mantener los valores de RAM como un arreglo de bytes
-    private byte[] data;
+    // Mantener los valores de RAM como un arreglo de ints
+    private int[] data;
 
     // La memoria mantiene el direccionamiento como una referencia al registro de direcciones de memoria.
     private IRegistro MAR;
@@ -22,19 +22,19 @@ public class Memoria {
 
     public Memoria(IRegistro MAR) {
         capacidadMemoria = 64; 
-        this.data = new byte[capacidadMemoria];
+        this.data = new int[capacidadMemoria];
         this.MAR = MAR;
         this.observers = new ArrayList<IRAMObserver>();
 
         // Cargar valores basura en la memoria (para probar)
         for (int i = 0; i < 16; i++) {
-            this.data[i] = (byte) ThreadLocalRandom.current().nextInt(0, 254);
+            this.data[i] = (int) ThreadLocalRandom.current().nextInt(0, 254);
         }
 
     }
 
     // Carga un valor en la dirección de memoria contenida en el registro de direcciones de memoria MAR
-    public void memoryIn(byte val) {
+    public void memoryIn(int val) {
         this.data[this.MAR.getValor()] = val;
         this.notifyObservers(this.MAR.getValor());
     }    
@@ -45,12 +45,12 @@ public class Memoria {
     }
 
     // Cambia manualmente una dirección de memoria (utilizada en WidgetRAM)
-    public void cambiarValor(int address, byte newVal) {
+    public void cambiarValor(int address, int newVal) {
         this.data[address] = newVal;
         this.notifyObservers(address);
     }
     // Devuleve el contenido de la memoria
-    public byte[] getData() {
+    public int[] getData() {
         return this.data;
     }
